requirements.txt
flask
flask_restful
Flask-SQLAlchemy
pandas
fastavro
sqlalchemy
psycopg2-binary